refactor: moving the kusto client
