+++ azurerm/internal/services/kusto/cilent/client.go
@@ -1,4 +1,4 @@
-package kusto
+package cilent
 
 import (
 	"github.com/Azure/azure-sdk-for-go/services/kusto/mgmt/2019-05-15/kusto"
@@ -11,7 +11,7 @@ type Client struct {
 	DataConnectionsClient *kusto.DataConnectionsClient
 }
 
-func BuildClient(o *common.ClientOptions) *Client {
+func NewClient(o *common.ClientOptions) *Client {
 	ClustersClient := kusto.NewClustersClientWithBaseURI(o.ResourceManagerEndpoint, o.SubscriptionId)
 	o.ConfigureClient(&ClustersClient.Client, o.ResourceManagerAuthorizer)
